+++ lib/generators/komponent/templates/styleguide/components/container/komponent_container_component.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module KomponentContainerComponent
+  extend ComponentHelper
+end