+++ lib/komponent.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-require "komponent/version"
+require 'komponent/version'
 
 module Komponent
-  require 'komponent/railtie' if defined?(Rails)
+  require 'komponent/engine' if defined?(Rails)
 end