+++ test/komponent/komponent_helper_test.rb
@@ -77,4 +77,31 @@ class KomponentHelperTest < ActionView::TestCase
       %(<div class="pong">Greetings from Ping</div>),
       component('pong').chomp
   end
+
+  def test_helper_lists_components
+    assert_equal(
+      [
+        'all',
+        'bar',
+        'foo',
+        'foo_bar',
+        'hello',
+        'ping',
+        'pong',
+        'required',
+        'world',
+      ],
+      components.keys
+    )
+  end
+
+  def test_helper_renders_with_doc
+    assert_equal \
+      %(<div class="all">🌎 😎</div>
+<pre class="komponent-code"><code>= component &quot;all&quot;, {
+  world: &quot;🌎&quot;,
+  sunglasses: &quot;😎&quot;
+}</code></pre>),
+      component_with_doc('all', world: "🌎", sunglasses: "😎").chomp
+  end
 end